Fix Windows build failure (#78)
Addresses #77

A behavior change was introduced such that either total memory or
limit/spike environment variables had to be set even if the config
parameter was passed. This fails for Windows as you cannot pass env vars
to the service.

To address, total or limit/spike is no longer required if config
parameter is addressed. Added a test to ensure this issue is not
reintroduced in the future.