Change default config server port to 55554
It appears Mac OS Big Sur uses port 55555:
https://developer.apple.com/forums/thread/671197. To ease user
experience, change the default port of the config server.

While this is a breaking change, the port is only exposed locally and
used by the support bundle script. As a result, this should not have
user impact beyond education of the updated port.