Cloud Foundry Buildpack for Collector
This is a deployment of the OpenTelemetry Collecter in the format of a
Cloud Foundry buildpack. This buildpack supplies the Collector to the
app it is used for. When the app is deployed it can run and configure
the Collector as a sidecar (as described in the README). This will allow
the Collector to observe the given app as well as the whole environment's
metrics.