check if monitor accepts endpoints before setting host/port (#1107)
Only select smart-agent monitors can be created via endpoints. This is
defined as a struct tag on the monitor configs "acceptsEndpoints". In
the case where we attempt to set Host/Port on a monitor that has it's
"acceptsEndpoints" value set to false, an error occurs and we do not
create the monitor.

This change would create the monitor but avoid setting the
Host/Port fields. This allows the receiver creator to
create any smart-agent monitor.