Enable Cloud Foundry receiver (#1414)
* Enable Cloud Foundry receiver

The Cloud Foundry receiver has been enabled in the OpenTelementry
Collector Contrib repository, and is now supported. This receiver
should also be supported in the Splunk distribution. This change
adds it as a dependency, adds it to the list of valid receivers,
and adds a reference to documentation, and the CHANGELOG.