Remove --mem-ballast-size-mib handling of arg (#2339)
The argument was removed from the core collector in v0.36.0. We kept supporting it in our distro but the support is broken. We never remove that argument as we do for others and it's being passed to the core collector in https://github.com/signalfx/splunk-otel-collector/blob/main/cmd/otelcol/main.go#L105 where it fails with the following error:

```
2022/12/08 16:25:34 main.go:106: application run finished with error: unknown flag: --mem-ballast-size-mib
```

So I don't think it makes sense to bring back the support of it