Fix exported module names to match directory locations
To refer to a specific version of a module from another repo, the Go
module system requires that the module path match the name of the
directory path inside the repo. See https://go.dev/ref/mod#vcs-dir

This requires changing the names of the exported modules to include
the "pkg" prefix used in the directory structure (and in the case of
smartagentextension, correcting the module name to reflect that it's
in the extension subdirectory, not the receiver subdirectory).

It also requires the name of the leaf directory match the module
name. This commit standardizes on the naming scheme used in the
OpenTelemetry Collector contrib repo, where subdirectories (and module
nams) include the type of the "plugin" as a suffix to the name of the
module and directory.

For example:
  processor/timestamp => pkg/processor/timestampprocessor
  receiver/smartagent => pkg/receiver/smartagentreceiver