[receiver/jmx] Add Docker support for JMX receiver
To properly support the jmx receiver in Docker, the dockerfile has
to download the JMX metrics gatherer, and have it available to
the collector. This change adds this functionality to the linux/macos
dockerfile, as well as the windows dockerfile.

Also, adds an integration test getting metrics from a docker container.