[discovery] Emit only one log record per matched endpoint
This is another step towards migrating from the discovery events to entity events. Entity batch has to have only one entity with a unique identifier.

There is no reason to send a log record for every match. The matching rules should not conflict with each other. We can document and assume that.

The receiving part (discovery config provider) don't need all the matching states. It only cares abut the latest state. And the successful state always wins. With this change, in case if we had a conflicting matching rules in both failed, partial and success sections, There is no reason to send log record per every would be unclear. However, given that we don't have any conflicting matching rules in the default discovery receiver configuration, it's safe to proceed.

As the next step, we will need to restructure the discovery rules schema to avoid any possible conflicts across different statuses.