[receiver/discovery] Remove `first_only`  field from the matching rule (#4593)
This is another step towards moving from discovery events to the entity events. The entity events will be emitted once a new matching rule matches a metric or log statement. After the initial emit the same entity state will be emitted on a regular basis, not when we get another match. This makes the `first_only` config option conflicting with the way how the entity events are expected to be emitted.

WIth this particular change, we switch behavior to `emit_first=true` by design. Given that all the discovery rules have the same value preconfigured, it's not a breaking change for users who rely on them. It is a breaking change for users who have custom discovery rules with emit_first=false. This is acceptable given the "In development" status of the receiver.