[discovery] Apply entity events schema the discovery receiver logs
This change updates the schema of the logs emitted by the discovery receiver to represent entity events. No information is dropped, but it was reshuffled to adopt for the new schema:
- Entity ID is made of one key/value pair moved from `discovery.endpoint.id` resource attribute
- All other resource attributes are moved to the entity state event attributes
- Log body was moved to the `discovery.event.message` entity state attribute

The Discovery config provider is updated to use the entity events.

The keys used in the entity ID and entity attributes are not finalized and are expected to be changed in the following PRs.