[receiver/discovery] Only track enpoint matching receiver rules
This way we can start tracking target endpoints in the correlation store and emit entity events on regular basis.

Also, this requires limiting discovered endpoint to only one receiver. If overlapping rule is found in more than one receiver, that endpoint will be ignored with a warning. If we keep allowing multiple receivers per endpoint it'll be hard to track successful discovery of endpoint/service because different receivers may provide conflicting results.