[receiver/discovery] Only track enpoint matching receiver rules (#4660)
This way, we can start tracking target endpoints in the correlation store and emit entity events on regular basis.

Also, this requires limiting discovered endpoint to only one receiver. If an overlapping rule is found in multiple receivers, that endpoint will be ignored with a warning. If we keep allowing multiple receivers per endpoint, it'll be unclear how to track successful discovery of endpoint/service because different receivers may provide conflicting results.

This change required pulling code from the receivercreator, which will be reused once the discovery receiver is upstreamed.