[receiver/discovery] Simplify config reporting (#4717)
This change removes the complicated logic of passing the encoded config by metrics receivers as part of a resource attribute, intercepting and patching them in the discovery receiver. Instead, we compute the embedded config on the fly for every emitted entity. In the short run, this hits the performance of the discovery receiver, but once we move the `metric.match` and `statement.match` discovery events to entity states in the following PRs, it won't be an issue.

Except for the performance implications, there are no functional changes in this PR.