[receiver/discovery] Refactor correlation store to be endpoint-based
Update all the methods to take endpointID as the first argument.

The most important change is repurposing of the attributes store. After https://github.com/signalfx/splunk-otel-collector/pull/4717, it's not being used anymore, but this will be used to store endpoint-specific attributes. So this change migrates the field from storing receiver's attributes to storing endpoint/correlation attributes.