[receiver/discovery] Assign correct resource in the metric evaluation
Metric evaluation emits entities with attribute from the first resource in the metric batch regardless where the matching metrics were found. This change fixes the bug. The resource attributes now assigned from the resource where the metrics are found.

It doesn't make sense to mention in the changelog because we don't have any pre-configured receiver that can run into this bug.