[receiver/discovery] Emit evaluation entities by the endpoint tracker (#4728)
This change consolidates the entity emitting in one place: endpoint tracker. Once the evaluation succeeds, the endpoint attributes are updated in the correlation store and the new entity state is emitted right away. Every subsequent state is emitted with all the added attributes.

The tests are kept with minimal changes to ensure that no attributes are lost in the emitted entities.

Currently, we emit superset of all the attributes that were previously emitted on regular basis or on the evaluation. Later, we will reduce the set and remove redundant attributes.