[receiver/discovery] Replace `log_record` field to `message`
`log_record` field isn't relevant anymore since we don't send discovery events as log records. The provided value is set to `discovery.message` entity attribute and sent to the backend on a regular basis. The additional `log_record.attributes` field is not being used. If needed in future, it can be added as another field in the statement section itself.