Avoid issues with receivers trying to update Splunk its dimensions
Currently, if a receiver has a dummy token it gets an error similar to this one:
Jul 04 14:25:19 christhianb-o11y2 otelcol[522]: 2024-07-04T14:25:19.870Z        
  error        dimensions/dimclient.go:247         Unable to update dimension, retrying       
  {"kind": "exporter", "data_type": "metrics", "name": "signalfx/internal", "error": 
  "error making HTTP request to https://api.us1.signalfx.com/v2/dimension/gcp_id/ps-1234/_/sfxagent: 
  Patch \"https://api.us1.signalfx.com/v2/dimension/...."

Once a valid token is configured it disappears. 
However, the purpose of configuring a Gateway would be to rely on this host to send everything to Splunk Observability.
Right after the proposed change is applied, the receiver's call starts using a valid token(Gateway Token) and the error disappears.