[chore] Upgrade core dependencies (#5136)
* [chore] Upgrade core dependencies

Update to 0.105.0 version + cherry pick open-telemetry/opentelemetry-collector#10716

Also, disabling both confmap upstream feature gates for now. Enabling confmap.unifyEnvVarExpansion doesn't have any effect other than introducing a bug. The warning messages that are expected to show with the FG are ignored in our repo

We have duplicated functionality in our distro which needs to be deprecated first. Then we can remove it and align with upstream.