[receiver/discovery] Fix metrics-only collection (#5359)
I missed some conditions in previous PR. If the receiver was configured on metrics pipeline only with some discovered services, it would block collector shutdown's. This change fixes that. It's not a typical use case to enable it on metrics pipeline only, so we don't need to test it extensively. I'll be covered to some extend in the following PRs.