[discovery] Move continuous discovery config parts to static files (#5381)
We cannot assume that ${SPLUNK_ACCESS_TOKEN} and ${SPLUNK_REALM} are aways set. It's not the case for the helm chart. We should move static config part of continuous discovery from discoverer to the config files. Having it in the `agent_config.yaml` has no overhead until the continuous discovery is enabled.