Stop supporting deprecated syntax for config source expansion
Expansion of bare config sources in the configuration is discontinued.

Strings like `$ENV` or `$include:/path/to/file.yaml` will no longer be expanded. Instead, use the `${env:ENV}` or `${include:/path/to/file.yaml}` syntax. There are only two symbols allowed after `$`: `{` and `$`. The collector will log an error and fail to start if it encounters a bare config source.