Remove ubuntu 16 from tests
Ubuntu 16 is only getting the extended paid security support from Cannonical these days. The tests are failing most of the time due to unreliable package endpoints for Ubuntu 16. We don't remove the support but remove want to make sure the CI is stable. So the tests for Ubuntu 16 are removed.